Move the RTT channel configuration into the RPC module
