+++ probe-rs-tools/src/bin/probe-rs/rpc/functions.rs
@@ -100,6 +100,7 @@ pub mod probe;
 pub mod reset;
 pub mod resume;
 pub mod rtt_client;
+pub mod rtt_config;
 pub mod stack_trace;
 pub mod test;
