+++ probe-rs-tools/src/bin/probe-rs/util/rtt.rs
@@ -1,4 +1,3 @@
-use clap::ValueEnum;
 use postcard_schema::Schema;
 use probe_rs::rtt::{self, DownChannel, Error, Rtt, UpChannel};
 use probe_rs::{Core, MemoryInterface};
@@ -7,42 +6,8 @@ use serde::{Deserialize, Serialize};
 pub(crate) mod client;
 pub(crate) mod processing;
 
+pub use crate::rpc::functions::rtt_config::{ChannelMode, DataFormat, RttChannelConfig};
 pub use processing::*;
-
-/// Used by serde to provide defaults for `RttChannelConfig::show_timestamps`
-fn default_show_timestamps() -> bool {
-    true
-}
-
-#[derive(
-    Debug, Copy, Clone, PartialEq, Eq, Default, docsplay::Display, Serialize, Deserialize, Schema,
-)]
-pub enum DataFormat {
-    #[default]
-    /// string
-    String,
-    /// binary
-    BinaryLE,
-    /// defmt
-    Defmt,
-}
-
-/// Specifies what to do when a channel doesn't have enough buffer space for a complete write on the
-/// target side.
-#[derive(Clone, Copy, Eq, PartialEq, Debug, Serialize, Deserialize, Schema, ValueEnum)]
-#[repr(u32)]
-pub enum ChannelMode {
-    /// Skip writing the data completely if it doesn't fit in its entirety.
-    NoBlockSkip = 0,
-
-    /// Write as much as possible of the data and ignore the rest.
-    NoBlockTrim = 1,
-
-    /// Block (spin) if the buffer is full. Note that if the application writes within a critical
-    /// section, using this mode can cause the application to freeze if the buffer becomes full and
-    /// is not read by the host.
-    BlockIfFull = 2,
-}
 
 impl From<ChannelMode> for rtt::ChannelMode {
     fn from(mode: ChannelMode) -> Self {
@@ -86,44 +51,6 @@ impl RttConfig {
             .iter()
             .find(|ch| ch.channel_number == Some(channel_number))
             .unwrap_or(&self.default_config)
-    }
-}
-
-/// The User specified configuration for each active RTT Channel.
-#[derive(Debug, Clone, Serialize, Deserialize, Schema)]
-#[serde(rename_all = "camelCase")]
-pub struct RttChannelConfig {
-    pub channel_number: Option<u32>,
-    #[serde(default)]
-    pub data_format: DataFormat,
-
-    /// RTT channel operating mode. Defaults to the target's configuration.
-    #[serde(default)]
-    pub mode: Option<ChannelMode>,
-
-    #[serde(default = "default_show_timestamps")]
-    /// Controls the inclusion of timestamps for [`DataFormat::String`] and [`DataFormat::Defmt`].
-    pub show_timestamps: bool,
-
-    #[serde(default)]
-    /// Controls the inclusion of source location information for DataFormat::Defmt.
-    pub show_location: bool,
-
-    #[serde(default)]
-    /// Controls the output format for DataFormat::Defmt.
-    pub log_format: Option<String>,
-}
-
-impl Default for RttChannelConfig {
-    fn default() -> Self {
-        RttChannelConfig {
-            channel_number: Default::default(),
-            data_format: Default::default(),
-            mode: Default::default(),
-            show_timestamps: default_show_timestamps(),
-            show_location: Default::default(),
-            log_format: Default::default(),
-        }
     }
 }
